requirement.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
backports.entry-points-selectable==1.1.1
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
datasets==2.0.0
dill==0.3.4
distlib==0.3.3
filelock==3.4.0
Flask==2.0.2
flatbuffers==2.0
frozenlist==1.3.0
fsspec==2022.2.0
gast==0.4.0
google-auth==2.4.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.6.0
huggingface-hub==0.4.0
idna==3.3
importlib-metadata==4.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
joblist==2.0.3
kaggle==1.5.12
keras==2.7.0
Keras-Preprocessing==1.1.2
libclang==12.0.0
Markdown==3.3.6
MarkupSafe==2.0.1
multidict==6.0.2
multiprocess==0.70.12.2
nltk==3.6.7
numpy==1.22.0
oauthlib==3.1.1
opendatasets==0.1.20
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.0
Pillow==9.0.1
platformdirs==2.4.0
protobuf==3.19.3
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dateutil==2.8.2
python-slugify==6.1.1
pytz==2021.3
PyYAML==6.0
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.0
responses==0.18.0
rsa==4.8
Sastrawi==1.0.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
text-unidecode==1.3
threadpoolctl==3.0.0
tqdm==4.62.3
tweet-preprocessor==0.6.0
typing_extensions==4.0.1
urllib3==1.26.8
virtualenv==20.10.0
w3lib==1.22.0
Werkzeug==2.0.2
wrapt==1.13.3
xxhash==3.0.0
yapf==0.32.0
yarl==1.7.2
zipp==3.7.0
